pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel", "setuptools_scm"]
build-backend = "setuptools.build_meta"

[tool.setuptools_scm]
normalize = true

[project]
name = "python-arango-async"
description = "Async Python Driver for ArangoDB"
authors = [
    {name= "Alexandru Petenchea", email = "alexandru.petenchea@arangodb.com" },
    {name = "Anthony Mahanna", email = "anthony.mahanna@arangodb.com"}
]
maintainers = [
    {name = "Alexandru Petenchea", email = "alexandru.petenchea@arangodb.com"},
    {name = "Anthony Mahanna", email = "anthony.mahanna@arangodb.com"}
]
keywords = ["arangodb", "python", "driver", "async"]
readme = "README.md"
dynamic = ["version"]
license = { file = "LICENSE" }
requires-python = ">=3.9"

classifiers = [
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Documentation :: Sphinx",
    "Typing :: Typed",
]

dependencies = [
    "packaging>=23.1",
    "setuptools>=42",
    "aiohttp>=3.9",
    "multidict>=6.0",
    "PyJWT>=2.8.0",
]

[tool.setuptools.dynamic]
version = { attr = "arangoasync.__version__" }

[project.optional-dependencies]
dev = [
    "black>=24.2",
    "flake8>=7.0",
    "isort>=5.10",
    "mypy>=1.10",
    "pre-commit>=3.7",
    "pytest>=8.2",
    "pytest-asyncio>=0.23.8",
    "pytest-cov>=5.0",
    "sphinx>=7.3",
    "sphinx_rtd_theme>=2.0",
    "types-setuptools",
]

[tool.setuptools.package-data]
"arangoasync" = ["py.typed"]

[project.urls]
homepage = "https://github.com/arangodb/python-arango-async"

[tool.setuptools]
packages = ["arangoasync"]


[tool.pytest.ini_options]
addopts = "-s -vv -p no:warnings"
minversion = "6.0"
testpaths = ["tests"]

[tool.coverage.run]
omit = [
    "arangoasync/version.py",
    "setup.py",
]

[tool.isort]
profile = "black"

[tool.mypy]
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
strict = true
